requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot==20.7
playwright
aiosqlite
python-dotenv
pydantic
